Day 24 optimized
This takes the runtime from 5-6s on part 2 to ~100ms by using a better data structure for lookups than trying to check if the Queue contains an existing state or not.

Now the slowest thing here is the GetNeighbors loop. I feel like that can be improved somehow, but I'm not sure if I can improve it and still retain its ergonomics or not. We shall see!