Make vectors readonly again
These changes to instantiate new vectors don't materially impact the runtime of either of these days, so it seems best to make the immutability explicit.

I _can_, however, make day 24 twice as fast by manipulating vectors in-place, but that's an ugly optimization that isn't actually necessary given its already-decent runtime (100ms -> 50ms) at the expense of making the code uglier and less DRY, so for now I'm not going to do that.